Refine PAC server
- Split the file operations to new PACDaemon
- More precise matching logic for handling PAC content http request
